feat: create LocalStorageFileSystem/GUI
Replaced and separated `Classzu.useLocalFileSystem` last 3 commits, and I think it is done.

But I haven't tested yet so I will test if this works.
Of cause not LOL. I will commit fixture changes from next comit.
