+++ src/LocalStorageFileSystem/GUI/index.ts
@@ -1,0 +1,3 @@
+import GUI from "./GUI";
+
+export default GUI;
