+++ src/LocalStorageFileSystem/LocalStorageFileSystem.ts
@@ -1,11 +1,12 @@
 import Classzu from "..";
 import * as ORM from './ORM'
+import GUI from './GUI'
 
 class LocalStorageFileSystem {
 
     public classzu: Classzu
     public ORM: typeof ORM = ORM
-
+    public GUI: GUI = new GUI(this)
     constructor(classzu: Classzu) {
 
         this.classzu = classzu;
@@ -13,6 +14,5 @@ class LocalStorageFileSystem {
     }
     
 }
-
 
 export default LocalStorageFileSystem;
